Rename ssr-paths
