+++ packages/pwa-kit-runtime/src/ssr/server/build-remote-server.js
@@ -44,7 +44,7 @@ import {
     getProxyPathBase,
     getHealthCheckPath,
     getSLASPrivateProxyPath
-} from '../../utils/ssr-paths'
+} from '../../utils/ssr-namespace-paths'
 import {applyProxyRequestHeaders} from '../../utils/ssr-server/configure-proxy'
 import awsServerlessExpress from 'aws-serverless-express'
 import expressLogging from 'morgan'
