+++ packages/pwa-kit-runtime/src/utils/ssr-namespace-paths.js
